requirements.txt
requests
django
djangorestframework
Celery
django-celery-beat
flower
plotly
